+++ src/js/vdlx-datagrid-column/attributes.js
@@ -75,12 +75,15 @@ export default [
     },
     {
         name: 'sort-by-formatted',
-        description: 'Use formatted values for sorting.'
+        description: 'Use formatted values for sorting. This is always set to enabled for calculated columns.',
+        valueType: 'boolean',
+        defaultValue: false
     },
     {
         name: 'filter-by-formatted',
         description:
-            'Use formatted values for filtering. This defaults to false unless the entity has a label in which case the default is true.'
+            'Use formatted values for filtering. This defaults to false unless the entity has a label in which case the default is true.',
+        valueType: 'boolean'
     },
     {
         name: 'editor-type',