+++ test/jest/mocks/insight-globals.js
@@ -4,19 +4,42 @@ const insightModules = {
     load: jest.fn()
 };
 const VDL = jest.fn();
-const insight = {};
+const insight = {
+    isDebugEnabled: jest.fn()
+};
 
-const dataUtilsSpy = jest.fn();
+const dataUtilsSpy = {
+    entityTypeIsNumber: jest.fn().mockReturnValue(false)
+};
 const createSparseDataSpy = jest.fn();
 const createDenseDataSpy = jest.fn();
 const autotableSelectOptionsSpy = jest.fn();
 const dialogsSpy = jest.fn().mockReturnValue({
     alert: jest.fn()
 });
-const enumsSpy = jest.fn().mockReturnValue({
-    DataType: {},
-    EntityManagementType: {}
-});
+const enumsMock = {
+    DataType: {
+        UNSUPPORTED: 'UNSUPPORTED',
+        BOOLEAN: 'BOOLEAN',
+        INTEGER: 'INTEGER',
+        REAL: 'REAL',
+        STRING: 'STRING',
+        SET: 'SET',
+        ARRAY: 'ARRAY',
+        DECISION_VARIABLE: 'DECISION_VARIABLE',
+        CONSTRAINT: 'CONSTRAINT',
+        MODEL: 'MODEL',
+        CONSTRAINT_TYPE: 'CONSTRAINT_TYPE',
+        VARIABLE_TYPE: 'VARIABLE_TYPE',
+        PROBLEM_STATUS: 'PROBLEM_STATUS'
+    },
+    EntityManagementType: {
+        DEFAULT: 'DEFAULT',
+        INPUT: 'INPUT',
+        RESULT: 'RESULT',
+        IGNORE: 'IGNORE'
+    }
+};
 const validatorFactorySpy = jest.fn();
 const insightGetterSpy = jest.fn().mockReturnValue(insight);
 const vdlValidatorRegistrySpy = jest.fn();
@@ -27,8 +50,8 @@ const modules = {
     'vdl/vdl-validator-factory': validatorFactorySpy,
     'dialogs': dialogsSpy,
     'insight-getter': insightGetterSpy,
-    'enums': enumsSpy,
     'external-libs/knockout': ko,
+    'enums': enumsMock,
     'components/table/create-sparse-data': createSparseDataSpy,
     'components/table/create-dense-data': createDenseDataSpy,
     'components/autotable-select-options': autotableSelectOptionsSpy